Rename hyperon.h to hyperon-gen.h
